backend/requirements.txt
# Core web framework
Flask==2.3.3
Flask-Cors==4.0.0

# Data processing and analysis
pandas==2.1.1
numpy==1.24.3
pandas_ta==0.3.14b0
scikit-learn==1.3.0
imbalanced-learn==0.11.0

# Deep Learning
tensorflow==2.15.0

# API and Data Fetching
alpha_vantage==2.3.1
newsapi-python==0.2.7
requests==2.31.0

# Sentiment Analysis
vaderSentiment==3.3.2

# Environment and configuration
python-dotenv==1.0.0
pytz==2023.3.post1

# Development tools
black==23.9.1
pylint==2.17.5